Support common layer of server part2 (#313)
